miniproject/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
twilio==8.10.0
requests==2.31.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.2
aiohttp==3.9.1
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.1.1
pyjwt==2.8.1
pydantic==2.5.0
email-validator==2.1.0





